refactoring of listOfSmilyeOptions/listOfTagOptions
