+++ pages/dashboard/AllEntries/[id].js
@@ -6,7 +6,12 @@ import HomeButton from "../../../components/HomeButton";
 import EditForm from "../../../components/EditForm";
 import { Heading } from "../../../components/Header/StyledHeader";
 
-const FormDetailsEdit = ({ formData, setFormData }) => {
+const FormDetailsEdit = ({
+  formData,
+  setFormData,
+  listOfSmileyOptions,
+  listOfTagOptions,
+}) => {
   const router = useRouter();
   const { id } = router.query;
   const entry = formData.find((element) => element.id === id);
@@ -14,7 +19,13 @@ const FormDetailsEdit = ({ formData, setFormData }) => {
   return (
     <>
       <Heading>Edit you Entry:</Heading>
-      <EditForm entry={entry} formData={formData} setFormData={setFormData} />
+      <EditForm
+        listOfSmileyOptions={listOfSmileyOptions}
+        listOfTagOptions={listOfTagOptions}
+        entry={entry}
+        formData={formData}
+        setFormData={setFormData}
+      />
       <Navbar>
         <BackButton />
         <AddButton />
