+++ pages/dashboard/AllEntries/index.js
@@ -10,69 +10,6 @@ import { motion } from "framer-motion";
 import PageLoader from "../../../components/PageLoader";
 import styled from "styled-components";
 
-const StyledSection = styled.section`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 48px;
-`;
-
-const StyledEntries = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 32px;
-  & > * {
-    order: -1;
-  }
-`;
-
-const StyledBatch = styled.div`
-  border: 1 px solid black;
-  background-color: white;
-  display: flex;
-  gap: 8px;
-  justify-content: center;
-  text-transform: uppercase;
-  align-items: center;
-  font-size: 16px;
-  font-weight: 600;
-  transition: background-color 0.2s ease-in-out;
-  background-color: rgba(140, 144, 122, 0.4);
-  margin: 10px 30px;
-  padding: 10px 16px;
-  border-radius: 999px;
-  border-bottom: 2px solid #8c907a;
-`;
-
-const StyledFilter = styled.div`
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 3vh;
-  z-index: 200;
-`;
-
-const StyledSelect = styled.select`
-  border-radius: 22px;
-  border: 1px solid #ececea;
-  padding: 2vw;
-  background-color: #d7b470;
-  width: 35vw;
-`;
-
-const listOfTagOptions = [
-  { tagName: "Family" },
-  { tagName: "Friends" },
-  { tagName: "Partner" },
-  { tagName: "Work" },
-  { tagName: "Hobby" },
-  { tagName: "Household" },
-  { tagName: "TV" },
-  { tagName: "Sports" },
-  { tagName: "Walk" },
-];
-
 export default function EntriesSection({
   filter,
   formData,
@@ -80,6 +17,8 @@ export default function EntriesSection({
   allEntriesCount,
   setEntries,
   entries,
+  listOfSmileyOptions,
+  listOfTagOptions,
 }) {
   const [selectedTag, setSelectedTag] = useState("");
   const tags = listOfTagOptions;
@@ -153,6 +92,8 @@ export default function EntriesSection({
                       setEntries={setEntries}
                       formData={formData}
                       setFormData={setFormData}
+                      listOfSmileyOptions={listOfSmileyOptions}
+                      listOfTagOptions={listOfTagOptions}
                     />
                   ))}
               </StyledEntries>
@@ -168,3 +109,54 @@ export default function EntriesSection({
     </>
   );
 }
+
+const StyledSection = styled.section`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 48px;
+`;
+
+const StyledEntries = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 32px;
+  & > * {
+    order: -1;
+  }
+`;
+
+const StyledBatch = styled.div`
+  border: 1 px solid black;
+  background-color: white;
+  display: flex;
+  gap: 8px;
+  justify-content: center;
+  text-transform: uppercase;
+  align-items: center;
+  font-size: 16px;
+  font-weight: 600;
+  transition: background-color 0.2s ease-in-out;
+  background-color: rgba(140, 144, 122, 0.4);
+  margin: 10px 30px;
+  padding: 10px 16px;
+  border-radius: 999px;
+  border-bottom: 2px solid #8c907a;
+`;
+
+const StyledFilter = styled.div`
+  display: flex;
+  justify-content: center;
+  position: sticky;
+  top: 3vh;
+  z-index: 200;
+`;
+
+const StyledSelect = styled.select`
+  border-radius: 22px;
+  border: 1px solid #ececea;
+  padding: 2vw;
+  background-color: #d7b470;
+  width: 35vw;
+`;
